+++ indico/modules/events/registration/fields/simple.py
@@ -6,6 +6,7 @@
 # LICENSE file for more details.
 
 from datetime import datetime
+from PIL import Image
 
 from marshmallow import ValidationError, fields, pre_load, validate, validates_schema
 
@@ -388,3 +389,24 @@ class EmailField(RegistrationFormFieldBase):
                 raise ValidationError(_('Invalid email address'))
 
         return _indico_email
+
+
+class PictureField(FileField):
+    name = 'picture'
+    setup_schema_fields = {
+        'min_picture_size': fields.Integer(load_default=0, validate=validate.Range(0, 1200)),
+    }
+
+    def get_validators(self, existing_registration):
+        def _picture_size_and_type(value):
+            min_picture_size = self.form_item.data.get('min_picture_size') or None
+            file = File.query.filter(File.uuid == value, ~File.claimed).first() if value else None
+            if file and file.content_type not in ('image/png', 'image/jpg', 'image/gif'):
+                raise ValidationError(_('This field can accept only .jpg, .png and .gif picture formats.'))
+            if file and min_picture_size:
+                with Image.open(file.open()) as picture:
+                    if min(picture.size) < min_picture_size:
+                        raise ValidationError(_('The uploaded picture pixels is smaller than the minimum size of {}.')
+                                              .format(min_picture_size))
+
+        return _picture_size_and_type